requirements.txt
python-dotenv
colorlog
requests
py-cord
gitpython
tzdata